1. new wxList code 2. fixes to allow compilation at -W4 with VisualC++ 6.0
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@1035 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/ole/droptgt.cpp
@@ -371,7 +371,7 @@ bool wxFileDropTarget::OnDrop(long x, long y, const void *pData)
   HDROP hdrop = (HDROP)pData;   // @@ it works, but I'm not sure about it
 
   // get number of files (magic value -1)
-  UINT nFiles = ::DragQueryFile(hdrop, -1, NULL, 0);
+  UINT nFiles = ::DragQueryFile(hdrop, (unsigned)-1, NULL, 0u);
   
   // for each file get the length, allocate memory and then get the name
   char **aszFiles = new char *[nFiles];
